+++ lib/themes/app_theme.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class AppTheme {
+  static ThemeData get theme => ThemeData(
+        colorScheme: ColorScheme.fromSeed(
+          seedColor: const Color(0xFFE60000),
+        ),
+        useMaterial3: true,
+        appBarTheme: const AppBarTheme(
+          backgroundColor: Color(0xFFE60000),
+          foregroundColor: Colors.white,
+        ),
+      );
+}
