+++ lib/views/home/home_screen.dart
@@ -1,0 +1,1 @@
+export '../../screens/home_screen.dart';
