+++ lib/views/paso_parametros/paso_parametros_screen.dart
@@ -1,0 +1,1 @@
+export '../../screens/details_screen.dart';
